src/Samples.Azure.Eventer.AppenderProcessor/requirements.txt
adal==1.2.7
aiohttp==3.8.1
aiosignal==1.2.0
applicationinsights==0.11.10
appnope==0.1.2
argon2-cffi==21.1.0
async-timeout==4.0.2
attrs==21.2.0
azure-common==1.1.27
azure-core==1.20.1
azure-eventhub==5.6.1
azure-eventhub-checkpointstoreblob-aio==1.1.4
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.2.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.3.0
azure-mgmt-resource==13.0.0
azure-mgmt-storage==11.2.0
azure-storage-blob==12.9.0
azureml-core==1.0.23
azureml-pipeline==1.0.23
azureml-pipeline-core==1.0.23
azureml-pipeline-steps==1.0.23
azureml-sdk==1.0.23
azureml-telemetry==1.0.23
azureml-train==1.0.23
azureml-train-core==1.0.23
azureml-train-restclients-hyperdrive==1.0.23
azureml-widgets==1.0.23
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
contextlib2==21.6.0
cryptography==3.4.8
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
docker==5.0.3
entrypoints==0.3
frozenlist==1.2.0
idna==3.3
ipykernel==6.5.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.0
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
jsonpickle==2.0.0
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
msrest==0.6.21
msrestazure==0.6.4
multidict==5.2.0
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
ndg-httpsclient==0.5.1
nest-asyncio==1.5.1
notebook==6.4.6
npTDMS==1.4.0
numpy==1.21.4
oauthlib==3.1.1
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
paramiko==2.8.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.12.0
prompt-toolkit==3.0.22
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.15.89
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.16.0
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.1
uamqp==1.5.0
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
widgetsnbextension==3.5.2
yarl==1.7.2
